Url info moved to right place
